Implements analysis json export function, an analysis class
